@@ -23,17 +23,15 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cloud_firestore: ^0.13.6
-  firebase_storage: ^3.1.6
-  provider: ^4.1.2
-  image_picker: ^0.6.7
-  adaptive_dialog: ^0.6.4
-  font_awesome_flutter: ^8.8.1
-  intl: ^0.16.0
+  provider: ^5.0.0
+  image_picker: ^0.7.5
+  adaptive_dialog: ^0.10.0
+  font_awesome_flutter: ^9.0.0
+  intl: ^0.17.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  json_annotation: ^3.0.0
+  cupertino_icons: ^1.0.3
+  json_annotation: ^4.0.1
   collection: ^1.14.13
   auto_size_text: ^2.1.0
   precept_script:
@@ -41,12 +39,12 @@ dependencies:
 
 
 dev_dependencies:
-  test: ^1.14.4
+  test: ^1.16.5
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.0
-  build_runner:
-  json_serializable: ^3.2.0
+  mockito: ^5.0.7
+  build_runner: ^2.0.2
+  json_serializable: ^4.1.1
   precept_mock_backend:
     path: ../precept_mock_backend
 
